requirements.txt
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
decorator==5.2.1
Django==5.1.6
djangorestframework==3.15.2
dotenv==0.9.9
executing==2.2.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
loguru==0.7.3
matplotlib-inline==0.1.7
mysqlclient==2.2.7
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.50
pure_eval==0.2.3
Pygments==2.19.1
python-dotenv==1.0.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
win32_setctime==1.2.0
